ref : 위치이동 감지하며 setPlaceList 해주는 useEffect kakaoMap 전역 적용
